Implement flag checking for pipelines
