+++ jwst/stpipe/pipeline.py
@@ -37,6 +37,7 @@ from . import config_parser
 from . import Step
 from . import crds_client
 from . import log
+from .step import get_disable_crds_steppars
 from ..lib.class_property import ClassInstanceMethod
 
 
@@ -171,8 +172,16 @@ class Pipeline(Step):
             The parameters as retrieved from CRDS. If there is an issue, log as such
             and return an empty config obj.
         """
+        pars_model = cls.get_pars_model()
         refcfg = ConfigObj()
         refcfg['steps'] = Section(refcfg, refcfg.depth + 1, refcfg.main, name="steps")
+
+        # Check if retrieval should be attempted.
+        if get_disable_crds_steppars():
+            log.log.debug(f'{pars_model.meta.reftype.upper()}: CRDS parameter reference retrieval disabled.')
+            return refcfg
+
+
         log.log.debug('Retrieving all substep parameters from CRDS')
         #
         # Iterate over the steps in the pipeline
@@ -183,7 +192,6 @@ class Pipeline(Step):
             )
         #
         # Now merge any config parameters from the step cfg file
-        pars_model = cls.get_pars_model()
         log.log.debug(f'Retrieving pipeline {pars_model.meta.reftype.upper()} parameters from CRDS')
         exceptions = crds_client.get_exceptions_module()
         try:
